+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.0.2] - 2019-02-08
+
+- fix "Maximum call stack" error when resolving js files after css files
+
 ## [2.0.1] - 2019-02-04
 
 - fix parsing `.tsx` files